Some more cleanup move decorators  and migrations to different file
